Standard Arguments

This documentation lists all the arguments in falcon that are used repeatedly by functions to output tables.

Arguments

adae

(data.frame)
dataset (typically ADAE) required to build table.

adex

(data.frame)
dataset (typically ADEX) required to build table.

advs

(data.frame)
dataset (typically ADVS) required to build table.

df

(data.frame)
dataset required to build table.

alt_counts_df

(character) alternative dataset (typically ADSL) used only to calculate column counts.

show_colcounts

(flag)
whether column counts should be printed.

arm_var

(character)
arm variable used to split table into columns.

id_var

(character)
variable used as unique subject identifier.

lbl_overall

(character)
if specified, an overall column will be added to the table with the given value as the column label.

vars

(vector of character)
variables from df to include in the table.

lbl_vars

(vector of character)
labels corresponding to variables in vars to print in the table. Labels should be ordered according to the order of variables in vars.

pref_var

(character)
preferred term variable from adae to include in the table.

lbl_pref_var

(character)
label corresponding to preferred term variable pref_var to print in the table.

.stats

(character)
statistics to include in the table. Includes statistics for all variable types (only the statistics that are valid for a given variable's type will be printed). See tern::summarize_vars() for options.

.formats

(named list of character)
list of formats corresponding to each value in .stats. Each name is a value in .stats and the corresponding value is the format that should be applied to that statistic. See formatters::list_valid_format_labels() for a list of valid formats.

na_rm

(flag)
whether NA levels should be removed from the table.

prune_0

(flag)
whether all-zero rows should be removed from the table.

annotations

(named list of character)
list of annotations to add to the table. Valid annotation types are title, subtitles, main_footer, and prov_footer. Each name-value pair should use the annotation type as name and the desired string as value.

na_level

(character)
string to represent missing values.