FDA Table 2: Baseline Demographic and Clinical Characteristics Safety Population, Pooled Analyses

FDA Table 2: Baseline Demographic and Clinical Characteristics Safety Population, Pooled Analyses

make_table_02(
  df,
  alt_counts_df = NULL,
  show_colcounts = TRUE,
  arm_var = "ARM",
  vars = c("SEX", "AGE", "AGEGR1", "RACE", "ETHNIC", "COUNTRY"),
  lbl_vars = formatters::var_labels(df, fill = TRUE)[vars],
  lbl_overall = "Total Population",
  .stats = c("mean_sd", "median_range", "count_fraction"),
  .formats = NULL,
  na_rm = FALSE,
  prune_0 = TRUE,
  annotations = NULL
)

Arguments

df

(data.frame)
dataset required to build table.

alt_counts_df

(character) alternative dataset (typically ADSL) used only to calculate column counts.

show_colcounts

(flag)
whether column counts should be printed.

arm_var

(character)
arm variable used to split table into columns.

vars

(vector of character)
variables from df to include in the table.

lbl_vars

(vector of character)
labels corresponding to variables in vars to print in the table. Labels should be ordered according to the order of variables in vars.

lbl_overall

(character)
if specified, an overall column will be added to the table with the given value as the column label.

.stats

(character)
statistics to include in the table. Includes statistics for all variable types (only the statistics that are valid for a given variable's type will be printed). See tern::summarize_vars() for options.

.formats

(named list of character)
list of formats corresponding to each value in .stats. Each name is a value in .stats and the corresponding value is the format that should be applied to that statistic. See formatters::list_valid_format_labels() for a list of valid formats.

na_rm

(flag)
whether NA levels should be removed from the table.

prune_0

(flag)
whether all-zero rows should be removed from the table.

annotations

(named list of character)
list of annotations to add to the table. Valid annotation types are title, subtitles, main_footer, and prov_footer. Each name-value pair should use the annotation type as name and the desired string as value.

Details

df must contain SAFFL and the variables specified by vars and arm_var.

If specified, alt_counts_df must contain SAFFL, USUBJID, and the variable specified by arm_var.

Flag variables (i.e. XXXFL) are expected to have two levels: "Y" (true) and "N" (false). Missing values in flag variables are treated as "N".

Columns are split by arm. Overall population column is included by default (see lbl_overall argument).

Information from either ADSUB or ADVS is generally included into df prior to analysis.

Numbers in table for non-numeric variables represent the absolute numbers of patients and fraction of N.

All-zero rows are removed by default (see prune_0 argument).

Examples

library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

adsl <- scda::synthetic_cdisc_dataset("rcd_2022_10_13", "adsl") %>%
  mutate(AGEGR1 = as.factor(case_when(
    AGE >= 17 & AGE < 65 ~ ">=17 to <65",
    AGE >= 65 ~ ">=65",
    AGE >= 65 & AGE < 75 ~ ">=65 to <75",
    AGE >= 75 ~ ">=75"
  )) %>% formatters::with_label("Age Group, years")) %>%
  formatters::var_relabel(
    AGE = "Age, years"
  )

tbl <- make_table_02(df = adsl)
tbl
#>                                                   A: Drug X            B: Placebo         C: Combination      Total Population 
#> Characteristic                                     (N=134)              (N=134)              (N=132)              (N=400)      
#> ———————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————
#> Sex                                                                                                                            
#>   F                                                79 (59%)            82 (61.2%)            70 (53%)           231 (57.8%)    
#>   M                                                55 (41%)            52 (38.8%)            62 (47%)           169 (42.2%)    
#> Age, years                                                                                                                     
#>   Mean (SD)                                       33.8 (6.6)           35.4 (7.9)           35.4 (7.7)           34.9 (7.4)    
#>   Median (Min - Max)                          33.0 (21.0 - 50.0)   35.0 (21.0 - 62.0)   35.0 (20.0 - 69.0)   34.0 (20.0 - 69.0)
#> Age Group, years                                                                                                               
#>   >=17 to <65                                     134 (100%)           134 (100%)          131 (99.2%)          399 (99.8%)    
#>   >=65                                                0                    0                 1 (0.8%)             1 (0.2%)     
#> Race                                                                                                                           
#>   ASIAN                                           68 (50.7%)            67 (50%)            73 (55.3%)           208 (52%)     
#>   BLACK OR AFRICAN AMERICAN                       31 (23.1%)           28 (20.9%)           32 (24.2%)           91 (22.8%)    
#>   WHITE                                           27 (20.1%)           26 (19.4%)           21 (15.9%)           74 (18.5%)    
#>   AMERICAN INDIAN OR ALASKA NATIVE                  8 (6%)             11 (8.2%)             6 (4.5%)            25 (6.2%)     
#>   MULTIPLE                                            0                 1 (0.7%)                0                 1 (0.2%)     
#>   NATIVE HAWAIIAN OR OTHER PACIFIC ISLANDER           0                 1 (0.7%)                0                 1 (0.2%)     
#> Ethnicity                                                                                                                      
#>   NOT REPORTED                                     6 (4.5%)            10 (7.5%)            11 (8.3%)            27 (6.8%)     
#>   HISPANIC OR LATINO                              15 (11.2%)           18 (13.4%)           15 (11.4%)            48 (12%)     
#>   NOT HISPANIC OR LATINO                         104 (77.6%)          103 (76.9%)          101 (76.5%)           308 (77%)     
#>   UNKNOWN                                          9 (6.7%)             3 (2.2%)             5 (3.8%)            17 (4.2%)     
#> Country                                                                                                                        
#>   CHN                                             74 (55.2%)           81 (60.4%)           64 (48.5%)          219 (54.8%)    
#>   USA                                             10 (7.5%)            13 (9.7%)            17 (12.9%)            40 (10%)     
#>   BRA                                             13 (9.7%)             7 (5.2%)            10 (7.6%)            30 (7.5%)     
#>   PAK                                              12 (9%)              9 (6.7%)            10 (7.6%)            31 (7.8%)     
#>   NGA                                               8 (6%)              7 (5.2%)            11 (8.3%)            26 (6.5%)     
#>   RUS                                              5 (3.7%)              8 (6%)              6 (4.5%)            19 (4.8%)     
#>   JPN                                              5 (3.7%)              4 (3%)              9 (6.8%)            18 (4.5%)     
#>   GBR                                               4 (3%)              3 (2.2%)             2 (1.5%)             9 (2.2%)     
#>   CAN                                              3 (2.2%)             2 (1.5%)             3 (2.3%)              8 (2%)