FDA Table 7: Deaths, Safety Population, Pooled Analyses

FDA Table 7: Deaths, Safety Population, Pooled Analyses

make_table_07(
  adae,
  alt_counts_df = NULL,
  show_colcounts = TRUE,
  arm_var = "ARM",
  lbl_overall = NULL,
  prune_0 = TRUE,
  na_level = "MISSING",
  annotations = NULL
)

Arguments

adae

(data.frame)
data (typically ADAE) required to build table.

alt_counts_df

(character) alternative (full) dataset (typically ADSL) used only to calculate column counts.

show_colcounts

(flag)
whether column counts should be printed.

arm_var

(character)
arm variable used to split table into columns.

lbl_overall

(character)
if specified, an overall column will be added to the table with the given value as the column label.

prune_0

(flag)
whether all-zero rows should be removed from the table.

na_level

(character)
string to represent missing values.

annotations

(named list of character)
list of annotations to add to the table. Valid annotations types are title, subtitles, main_footer, and prov_footer. Each name-value pair should use the annotation type as name and the desired string as value.

Details

adae must contain SAFFL, USUBJID, TRTEMFL, DTHFL, DTHCAUS, and the variable specified by arm_var.

If specified, alt_counts_df must contain SAFFL, USUBJID, and the variable specified by arm_var.

Flag variables (i.e. XXXFL) are expected to have two levels: "Y" (true) and "N" (false). Missing values in flag variables are treated as "N".

Columns are split by arm. Overall population column is excluded by default (see lbl_overall argument).

Numbers in table represent the absolute numbers of patients and fraction of N for category summary rows and fraction of n (number of patients in current category) for all other rows.

All-zero rows are removed by default (see prune_0 argument).

Examples

adsl <- scda::synthetic_cdisc_dataset("rcd_2022_10_13", "adsl")
adae <- scda::synthetic_cdisc_dataset("rcd_2022_10_13", "adae")

set.seed(1)
adae$TRTEMFL <- ifelse(adae$USUBJID %in% sample(adsl$USUBJID, size = as.integer(nrow(adsl) / 3)), "N", "Y")

tbl <- make_table_07(adae = adae, alt_counts_df = adsl)
tbl
#>                                   A: Drug X    B: Placebo   C: Combination
#> Deaths                             (N=134)      (N=134)        (N=132)    
#> ——————————————————————————————————————————————————————————————————————————
#> Total deaths                      24 (17.9%)   19 (14.2%)     20 (15.2%)  
#>   ADVERSE EVENT                   8 (33.3%)    6 (31.6%)      10 (50.0%)  
#>   DISEASE PROGRESSION             8 (33.3%)    5 (26.3%)      6 (30.0%)   
#>   LOST TO FOLLOW UP                2 (8.3%)     1 (5.3%)       1 (5.0%)   
#>   MISSING                          2 (8.3%)    3 (15.8%)       1 (5.0%)   
#>   Post-study reporting of death    1 (4.2%)     1 (5.3%)       1 (5.0%)   
#>   SUICIDE                          2 (8.3%)    2 (10.5%)       1 (5.0%)   
#>   UNKNOWN                          1 (4.2%)     1 (5.3%)          0       
#> Treatment-emergent deaths         12 (9.0%)    15 (11.2%)     16 (12.1%)  
#>   ADVERSE EVENT                   3 (25.0%)    5 (33.3%)      9 (56.2%)   
#>   DISEASE PROGRESSION             3 (25.0%)    5 (33.3%)      4 (25.0%)   
#>   LOST TO FOLLOW UP                1 (8.3%)        0              0       
#>   MISSING                         2 (16.7%)     1 (6.7%)       1 (6.2%)   
#>   Post-study reporting of death       0         1 (6.7%)       1 (6.2%)   
#>   SUICIDE                         2 (16.7%)    2 (13.3%)       1 (6.2%)   
#>   UNKNOWN                          1 (8.3%)     1 (6.7%)          0       
#> Nontreatment-emergent deaths      12 (9.0%)     4 (3.0%)       4 (3.0%)   
#>   ADVERSE EVENT                   5 (41.7%)    1 (25.0%)      1 (25.0%)   
#>   DISEASE PROGRESSION             5 (41.7%)        0          2 (50.0%)   
#>   LOST TO FOLLOW UP                1 (8.3%)    1 (25.0%)      1 (25.0%)   
#>   MISSING                             0        2 (50.0%)          0       
#>   Post-study reporting of death    1 (8.3%)        0              0