FDA Table 8: All Individual Patient Deaths, Safety Population, Pooled Analyses

FDA Table 8: All Individual Patient Deaths, Safety Population, Pooled Analyses

make_table_08(
  adae,
  adex,
  arm_var = "ARM",
  dth_vars = c("DTHCAUS", "DTHCAT"),
  lbl_dth_vars = c("Cause of Death\nMedDRA\nPreferred Term",
    "Cause of Death\nVerbatim Term"),
  na_level = "NA",
  annotations = NULL
)

Arguments

adae

(data.frame)
dataset (typically ADAE) required to build table.

adex

(data.frame)
dataset (typically ADEX) required to build table.

arm_var

(character)
arm variable used to split table into columns.

dth_vars

(vector of character)
additional death variables from adae to include in the table.

lbl_dth_vars

(vector of character)
labels corresponding to variables in dth_vars to print in the table. Labels should be ordered according to the order of variables in dth_vars.

na_level

(character)
string to represent missing values.

annotations

(named list of character)
list of annotations to add to the table. Valid annotation types are title, subtitles, main_footer, and prov_footer. Each name-value pair should use the annotation type as name and the desired string as value.

Details

adae must contain SAFFL, USUBJID, AGE, SEX, AESDTH, DTHADY, and the variables specified by dth_vars and arm_var.

adex must contain SAFFL, USUBJID, PARAMCD, TRTSDT, TRTEDT, AVAL, and AVALU.

Flag variables (i.e. XXXFL) are expected to have two levels: "Y" (true) and "N" (false). Missing values in flag variables are treated as "N".

Examples

adae <- scda::synthetic_cdisc_dataset("rcd_2022_10_13", "adae")
adex <- scda::synthetic_cdisc_dataset("rcd_2022_10_13", "adex")

tbl <- make_table_08(adae = adae, adex = adex)
head(tbl, 20)
#>                                                         Dosing    Study      Cause of Death                         
#>                                      Age/              Duration   Day of         MedDRA            Cause of Death   
#> Study Arm        Patient ID         Gender   Dosage     (Days)    Death      Preferred Term         Verbatim Term   
#> ————————————————————————————————————————————————————————————————————————————————————————————————————————————————————
#> A: Drug X   AB12345-BRA-1-id-42     36/M     5040 mg   37         NA               NA                    NA         
#>             AB12345-BRA-1-id-93     34/F     6720 mg   37         658         ADVERSE EVENT         ADVERSE EVENT   
#>             AB12345-BRA-11-id-345   37/F     7680 mg   37         NA               NA                    NA         
#>             AB12345-BRA-11-id-397   38/M     7680 mg   37         NA               NA                    NA         
#>             AB12345-BRA-13-id-177   24/M     4800 mg   37         NA               NA                    NA         
#>             AB12345-BRA-15-id-36    38/F     7200 mg   37         813      DISEASE PROGRESSION   PROGRESSIVE DISEASE
#>             AB12345-BRA-2-id-296    44/F     4800 mg   37         NA               NA                    NA         
#>             AB12345-CAN-11-id-139   31/M     6000 mg   37         564         ADVERSE EVENT         ADVERSE EVENT   
#>             AB12345-CHN-1-id-123    27/F     5280 mg   37         751      DISEASE PROGRESSION   PROGRESSIVE DISEASE
#>             AB12345-CHN-1-id-133    25/F     5520 mg   37         NA               NA                    NA         
#>             AB12345-CHN-1-id-235    39/M     5520 mg   37         969         ADVERSE EVENT         ADVERSE EVENT   
#>             AB12345-CHN-1-id-275    27/M     7200 mg   37         NA               NA                    NA         
#>             AB12345-CHN-1-id-277    41/M     8160 mg   37         NA               NA                    NA         
#>             AB12345-CHN-1-id-315    32/M     6480 mg   37         NA               NA                    NA         
#>             AB12345-CHN-1-id-316    25/F     8400 mg   37         NA               NA                    NA         
#>             AB12345-CHN-1-id-329    48/M     5520 mg   37         NA               NA                    NA         
#>             AB12345-CHN-1-id-47     25/F     5520 mg   37         NA               NA                    NA         
#>             AB12345-CHN-1-id-53     48/M     6960 mg   37         NA               NA                    NA         
#>             AB12345-CHN-1-id-69     32/F     6960 mg   37         NA               NA                    NA         
#>             AB12345-CHN-1-id-71     33/M     7440 mg   37         NA               NA                    NA