FDA Table 13: Patients With Common Adverse Events Occurring at >=XX% Frequency, Safety Population, Pooled Analyses

FDA Table 13: Patients With Common Adverse Events Occurring at >=XX% Frequency, Safety Population, Pooled Analyses

make_table_13(
  adae,
  alt_counts_df = NULL,
  show_colcounts = TRUE,
  min_freq = 0.05,
  arm_var = "ARM",
  pref_var = "AETERM",
  lbl_pref_var = formatters::var_labels(adae, fill = TRUE)[pref_var],
  lbl_overall = NULL,
  annotations = NULL
)

Arguments

adae

(data.frame)
data (typically ADAE) required to build table.

alt_counts_df

(character) alternative (full) dataset (typically ADSL) used only to calculate column counts.

show_colcounts

(flag)
whether column counts should be printed.

min_freq

(proportion)
minimum % frequency (fraction of N) required in any column to include each row.

arm_var

(character)
arm variable used to split table into columns.

pref_var

(character)
preferred term variable from adae to include in the table.

lbl_pref_var

(character)
label corresponding to preferred term variable pref_var to print in the table.

lbl_overall

(character)
if specified, an overall column will be added to the table with the given value as the column label.

annotations

(named list of character)
list of annotations to add to the table. Valid annotations types are title, subtitles, main_footer, and prov_footer. Each name-value pair should use the annotation type as name and the desired string as value.

Details

adae must contain SAFFL, USUBJID, and the variables specified by pref_var and arm_var.

If specified, alt_counts_df must contain variables SAFFL and USUBJID.

Flag variables (i.e. XXXFL) are expected to have two levels: "Y" (true) and "N" (false). Missing values in flag variables are treated as "N".

Columns are split by arm. Overall population column is excluded by default (see lbl_overall argument).

Numbers in table represent the absolute numbers of patients and fraction of N.

Examples

adsl <- scda::synthetic_cdisc_dataset("rcd_2022_10_13", "adsl")
adae <- scda::synthetic_cdisc_dataset("rcd_2022_10_13", "adae")

tbl <- make_table_13(adae = adae, alt_counts_df = adsl)
tbl
#>                                       A: Drug X    B: Placebo   C: Combination
#> Reported Term for the Adverse Event    (N=134)      (N=134)        (N=132)    
#> ——————————————————————————————————————————————————————————————————————————————
#> trm A.1.1.1.1                         50 (37.3%)   45 (33.6%)     63 (47.7%)  
#> trm A.1.1.1.2                         48 (35.8%)   48 (35.8%)     50 (37.9%)  
#> trm B.1.1.1.1                         47 (35.1%)   49 (36.6%)     43 (32.6%)  
#> trm B.2.1.2.1                         49 (36.6%)   44 (32.8%)     52 (39.4%)  
#> trm B.2.2.3.1                         48 (35.8%)   54 (40.3%)     51 (38.6%)  
#> trm C.1.1.1.3                         43 (32.1%)   46 (34.3%)     43 (32.6%)  
#> trm C.2.1.2.1                         35 (26.1%)   48 (35.8%)     55 (41.7%)  
#> trm D.1.1.1.1                         50 (37.3%)   42 (31.3%)     51 (38.6%)  
#> trm D.1.1.4.2                         48 (35.8%)   42 (31.3%)     50 (37.9%)  
#> trm D.2.1.5.3                         47 (35.1%)   58 (43.3%)     57 (43.2%)