FDA Table 6: Overview of Adverse Events, Safety Population, Pooled Analyses

FDA Table 6: Overview of Adverse Events, Safety Population, Pooled Analyses

make_table_14(
  adae,
  alt_counts_df = NULL,
  show_colcounts = TRUE,
  arm_var = "ARM",
  fmqsc_var = "FMQ01SC",
  fmqnam_var = "FMQ01NAM",
  lbl_overall = NULL,
  prune_0 = FALSE,
  na_level = "<Missing>",
  annotations = NULL
)

Arguments

adae

(data.frame)
dataset (typically ADAE) required to build table.

alt_counts_df

(character) alternative dataset (typically ADSL) used only to calculate column counts.

show_colcounts

(flag)
whether column counts should be printed.

arm_var

(character)
arm variable used to split table into columns.

fmqsc_var

(character)
FMQ scope variable to use in table.

fmqnam_var

(character)
FMQ reference name variable to use in table.

lbl_overall

(character)
if specified, an overall column will be added to the table with the given value as the column label.

prune_0

(flag)
whether all-zero rows should be removed from the table.

na_level

(character)
string to represent missing values.

annotations

(named list of character)
list of annotations to add to the table. Valid annotation types are title, subtitles, main_footer, and prov_footer. Each name-value pair should use the annotation type as name and the desired string as value.

Details

adae must contain the variables SAFFL, USUBJID, AEBODSYS, and the variables specified by arm_var, fmqsc_var, and fmqnam_var.

If specified, alt_counts_df must contain variables SAFFL and USUBJID.

Flag variables (i.e. XXXFL) are expected to have two levels: "Y" (true) and "N" (false). Missing values in flag variables are treated as "N".

Columns are split by arm. Overall population column is excluded by default (see lbl_overall argument).

Numbers in table represent the absolute numbers of patients and fraction of N.

All-zero rows are not removed by default (see prune_0 argument).

Examples

adae <- scda::synthetic_cdisc_dataset("rcd_2022_10_13", "adae")
adae <- dplyr::rename(adae, FMQ01SC = SMQ01SC, FMQ01NAM = SMQ01NAM)
levels(adae$FMQ01SC) <- c("BROAD", "NARROW")
adae$FMQ01SC[is.na(adae$FMQ01SC)] <- "NARROW"

tbl <- make_table_14(adae = adae)
tbl
#> Body System or Organ Class                    BROAD                                      NARROW                 
#>   FMQ                        A: Drug X    B: Placebo   C: Combination   A: Drug X    B: Placebo   C: Combination
#>                               (N=119)      (N=139)        (N=141)        (N=490)      (N=483)        (N=562)    
#> ————————————————————————————————————————————————————————————————————————————————————————————————————————————————
#> cl A.1                                                                                                          
#>   C.1.1.1.3/B.2.2.3.1 AESI       0            0              0              0            0              0       
#>   <Missing>                      0            0              0          78 (15.9%)   75 (15.5%)     89 (15.8%)  
#> cl B.1                                                                                                          
#>   C.1.1.1.3/B.2.2.3.1 AESI       0            0              0              0            0              0       
#>   <Missing>                      0            0              0          47 (9.6%)    49 (10.1%)     43 (7.7%)   
#> cl B.2                                                                                                          
#>   C.1.1.1.3/B.2.2.3.1 AESI   48 (40.3%)   54 (38.8%)     51 (36.2%)         0            0              0       
#>   <Missing>                      0            0              0          49 (10.0%)   44 (9.1%)      52 (9.3%)   
#> cl C.1                                                                                                          
#>   C.1.1.1.3/B.2.2.3.1 AESI   43 (36.1%)   46 (33.1%)     43 (30.5%)         0            0              0       
#>   <Missing>                      0            0              0              0            0              0       
#> cl C.2                                                                                                          
#>   C.1.1.1.3/B.2.2.3.1 AESI       0            0              0              0            0              0       
#>   <Missing>                      0            0              0          35 (7.1%)    48 (9.9%)      55 (9.8%)   
#> cl D.1                                                                                                          
#>   C.1.1.1.3/B.2.2.3.1 AESI       0            0              0              0            0              0       
#>   <Missing>                      0            0              0          79 (16.1%)   67 (13.9%)     80 (14.2%)  
#> cl D.2                                                                                                          
#>   C.1.1.1.3/B.2.2.3.1 AESI       0            0              0              0            0              0       
#>   <Missing>                      0            0              0          47 (9.6%)    58 (12.0%)     57 (10.1%)