FDA Table 20: Adverse Events of Special Interest Assessment, Safety Population, Pooled Analysis (or Trial X)

FDA Table 20: Adverse Events of Special Interest Assessment, Safety Population, Pooled Analysis (or Trial X)

make_table_20(
  adae,
  alt_counts_df = NULL,
  show_colcounts = TRUE,
  arm_var = "ARM",
  pref_var = "AEDECOD",
  aesifl_var = "AESIFL",
  aelabfl_var = "AELABFL",
  lbl_overall = NULL,
  prune_0 = TRUE,
  annotations = NULL
)

Arguments

adae

(data.frame)
dataset (typically ADAE) required to build table.

alt_counts_df

(character) alternative dataset (typically ADSL) used only to calculate column counts.

show_colcounts

(flag)
whether column counts should be printed.

arm_var

(character)
arm variable used to split table into columns.

pref_var

(character)
preferred term variable from adae to include in the table.

aesifl_var

(character)
variable from adae that indicates adverse events of special interest.

aelabfl_var

(character)
variable from adae that indicates a laboratory assessment.

lbl_overall

(character)
if specified, an overall column will be added to the table with the given value as the column label.

prune_0

(flag)
whether all-zero rows should be removed from the table.

annotations

(named list of character)
list of annotations to add to the table. Valid annotation types are title, subtitles, main_footer, and prov_footer. Each name-value pair should use the annotation type as name and the desired string as value.

Details

adae must contain SAFFL, USUBJID, AEDECOD, AESEV, AESER, AESDTH, EOSSTT, AEREL, and the variables specified by pref_var, aesifl_var, aelabfl_var, and arm_var.

If specified, alt_counts_df must contain SAFFL, USUBJID, and the variable specified by arm_var.

Flag variables (i.e. XXXFL) are expected to have two levels: "Y" (true) and "N" (false). Missing values in flag variables are treated as "N".

Columns are split by arm. Overall population column is excluded by default (see lbl_overall argument).

Numbers in table represent the absolute numbers of patients and fraction of N.

All-zero rows are removed by default (see prune_0 argument).

Examples

adsl <- scda::synthetic_cdisc_dataset("rcd_2022_10_13", "adsl")
adae <- scda::synthetic_cdisc_dataset("rcd_2022_10_13", "adae")

set.seed(1)
adae$AESIFL <- ifelse(adae$AESOC %in% c("cl A", "cl D"), "Y", "N")
adae$AELABFL <- sample(c("Y", "N"), nrow(adae), replace = TRUE)

tbl <- make_table_20(adae = adae, alt_counts_df = adsl)
tbl
#>                                 A: Drug X    B: Placebo    C: Combination
#> AESI Assessment                  (N=134)       (N=134)        (N=132)    
#> —————————————————————————————————————————————————————————————————————————
#> AE grouping related to AESI    107 (79.9%)   112 (83.6%)    113 (85.6%)  
#>   dcd A.1.1.1.1                50 (37.3%)    45 (33.6%)      63 (47.7%)  
#>   dcd A.1.1.1.2                48 (35.8%)    48 (35.8%)      50 (37.9%)  
#>   dcd D.1.1.1.1                50 (37.3%)    42 (31.3%)      51 (38.6%)  
#>   dcd D.1.1.4.2                48 (35.8%)    42 (31.3%)      50 (37.9%)  
#>   dcd D.2.1.5.3                47 (35.1%)    58 (43.3%)      57 (43.2%)  
#> Maximum severity                                                         
#>   MILD                         16 (11.9%)    19 (14.2%)      17 (12.9%)  
#>   MODERATE                     41 (30.6%)    51 (38.1%)      45 (34.1%)  
#>   SEVERE                       50 (37.3%)    42 (31.3%)      51 (38.6%)  
#> Serious                        75 (70.1%)    77 (68.8%)      83 (73.5%)  
#>   Deaths                       50 (46.7%)    42 (37.5%)      51 (45.1%)  
#> Resulting in discontinuation   35 (32.7%)    34 (30.4%)      33 (29.2%)  
#> Relatedness                    74 (69.2%)    74 (66.1%)      84 (74.3%)  
#> Laboratory Assessment          89 (66.4%)    86 (64.2%)      90 (68.2%)